refactor(cli): extract configuration logic to dedicated module
- Move project root resolution and YAML config parsing from cli.py
  to apps_platform/config.py to decouple core logic from CLI.
- Update documentation to reflect apps_platform directory structure.
- Simplify install.sh by removing redundant package checks and links.
- Delete obsolete validation script and changelog file.
